+++ CF/1253/D/CF.1253D.0.cpp
@@ -1,15 +1,15 @@
 #include<cstdio>
 using namespace std;
-long f[200001];
-long find(long p)
+int f[200001];
+int find(int p)
 {
 	if(f[p]!=p)
 		f[p]=find(f[p]);
 	return f[p];
 }
-void merge(long a,long b)
+void merge(int a,int b)
 {
-	long fa=find(a),fb=find(b);
+	int fa=find(a),fb=find(b);
 	if(fa>fb)// Make sure the root point has the maximum index
 		f[fb]=fa;
 	else
@@ -17,17 +17,17 @@ void merge(long a,long b)
 }
 int main()
 {
-	long n,m;
-	scanf("%ld%ld",&n,&m);
-	for(long i=0;i<=n;++i)
+	int n,m;
+	scanf("%d%d",&n,&m);
+	for(int i=0;i<=n;++i)
 		f[i]=i;
-	for(long i=0,x,y;i<m;++i)
+	for(int i=0,x,y;i<m;++i)
 	{
-		scanf("%ld%ld",&x,&y);
+		scanf("%d%d",&x,&y);
 		merge(x-1,y-1);
 	}
-	long ans=0;
-	for(long i=0,p=find(0);i<n;++i)// Scan the sequence and add edges if necessary, p is the maximum index in this block
+	int ans=0;
+	for(int i=0,p=find(0);i<n;++i)// Scan the sequence and add edges if necessary, p is the maximum index in this block
 	{
 		if(find(i)!=p)// Find a point which does not match
 		{
@@ -38,6 +38,6 @@ int main()
 		if(i==p && i<n-1)// Jump to the next connected block
 			p=find(i+1);
 	}
-	printf("%ld",ans);
+	printf("%d",ans);
 	return 0;
 }
