+++ CF/1257/B/CF.1257B.0.cpp
@@ -5,8 +5,8 @@ int main()
 	int t;
 	for(scanf("%d",&t);t>0;--t)
 	{
-		long x,y;
-		scanf("%ld%ld",&x,&y);
+		int x,y;
+		scanf("%d%d",&x,&y);
 		printf((x>=y || x>3 || x==2 && y==3)?"YES\n":"NO\n");
 	}
 	return 0;
