+++ CF/1257/E/CF.1257E.0.cpp
@@ -1,9 +1,9 @@
 #include<cstdio>
 using namespace std;
-long num[200000],l[200000];
-void qsort(long l,long r)// Sort the array num in increasing order
+int num[200000],l[200000];
+void qsort(int l,int r)// Sort the array num in increasing order
 {
-	long x=l,y=r,std=num[(l+r)/2];
+	int x=l,y=r,std=num[(l+r)/2];
 	while(x<y)
 	{
 		while(num[x]<std)
@@ -12,7 +12,7 @@ void qsort(long l,long r)// Sort the array num in increasing order
 			--y;
 		if(x<=y)
 		{
-			long tmp=num[x];
+			int tmp=num[x];
 			num[x++]=num[y];
 			num[y--]=tmp;
 		}
@@ -22,7 +22,7 @@ void qsort(long l,long r)// Sort the array num in increasing order
 	if(x<r)
 		qsort(x,r);
 }
-long find(long a,long b,long p)// Find the position to be replaced
+int find(int a,int b,int p)// Find the position to be replaced
 {
 	if(a>b)
 		return a;
@@ -30,20 +30,20 @@ long find(long a,long b,long p)// Find the position to be replaced
 }
 int main()
 {
-	long a,b,c;
-	scanf("%ld%ld%ld",&a,&b,&c);
-	for(long i=0;i<a+b+c;++i)
-		scanf("%ld",&num[i]);
+	int a,b,c;
+	scanf("%d%d%d",&a,&b,&c);
+	for(int i=0;i<a+b+c;++i)
+		scanf("%d",&num[i]);
 	qsort(0,a-1);// Sort everyone's problems to make sure their problems are increasing
 	qsort(a,a+b-1);
 	qsort(a+b,a+b+c-1);
-	long len=1;// Searching for the longest increasing sequence
+	int len=1;// Searching for the longest increasing sequence
 	l[0]=num[0];
 	for(int i=1;i<a+b+c;++i)
 		if(num[i]>l[len-1])
 			l[len++]=num[i];
 		else
 			l[find(0,len,num[i])]=num[i];
-	printf("%ld",a+b+c-len);
+	printf("%d",a+b+c-len);
 	return 0;
 }
