+++ CF/1296/D/CF.1296D.0.cpp
@@ -1,9 +1,9 @@
 #include<cstdio>
 using namespace std;
-long cost[200000];
-void qsort(long l,long r)// Sort the cost array in increasing order
+int cost[200000];
+void qsort(int l,int r)// Sort the cost array in increasing order
 {
-	long x=l,y=r,std=cost[(l+r)/2];
+	int x=l,y=r,std=cost[(l+r)/2];
 	while(x<y)
 	{
 		while(cost[x]<std)
@@ -12,7 +12,7 @@ void qsort(long l,long r)// Sort the cost array in increasing order
 			--y;
 		if(x<=y)
 		{
-			long tmp=cost[x];
+			int tmp=cost[x];
 			cost[x++]=cost[y];
 			cost[y--]=tmp;
 		}
@@ -24,17 +24,17 @@ void qsort(long l,long r)// Sort the cost array in increasing order
 }
 int main()
 {
-	long n,a,b,k;
-	scanf("%ld%ld%ld%ld",&n,&a,&b,&k);
-	for(long i=0,h;i<n;++i)// Calculate the cost
+	int n,a,b,k;
+	scanf("%d%d%d%d",&n,&a,&b,&k);
+	for(int i=0,h;i<n;++i)// Calculate the cost
 	{
-		scanf("%ld",&h);
+		scanf("%d",&h);
 		h=(h-1)%(a+b)+1;
 		cost[i]=(h<=a)?0:(h-1)/a;
 	}
 	qsort(0,n-1);// Sort the cost
-	long ans=0;
-	for(long i=0;i<n && k>=cost[i];++i)// Choose greedily
+	int ans=0;
+	for(int i=0;i<n && k>=cost[i];++i)// Choose greedily
 	{
 		k-=cost[i];
 		++ans;
