+++ CF/1304/A/CF.1304A.0.cpp
@@ -5,9 +5,9 @@ int main()
 	int t;
 	for(scanf("%d",&t);t>0;--t)
 	{
-		long x,y,a,b;
-		scanf("%ld%ld%ld%ld",&x,&y,&a,&b);
-		printf("%ld\n",(y>=x && (y-x)%(a+b)==0)?(y-x)/(a+b):-1);
+		int x,y,a,b;
+		scanf("%d%d%d%d",&x,&y,&a,&b);
+		printf("%d\n",(y>=x && (y-x)%(a+b)==0)?(y-x)/(a+b):-1);
 	}
 	return 0;
 }
