+++ CF/1313/B/CF.1313B.0.cpp
@@ -5,12 +5,12 @@ int main()
 	int t;
 	for(scanf("%d",&t);t>0;--t)
 	{
-		long n,x,y;
-		scanf("%ld%ld%ld",&n,&x,&y);
+		int n,x,y;
+		scanf("%d%d%d",&n,&x,&y);
 		if(x+y>n)
-			printf("%ld %ld\n",(x+y==2*n)?n:x+y+1-n,n);
+			printf("%d %d\n",(x+y==2*n)?n:x+y+1-n,n);
 		else
-			printf("%ld %ld\n",1,x+y-1);
+			printf("%d %d\n",1,x+y-1);
 	}
 	return 0;
 }
