+++ CF/1323/C/CF.1323C.0.cpp
@@ -3,7 +3,7 @@ using namespace std;
 char s[1000001];
 int main()
 {
-	long n,cnt=0,ans=0;// cnt is the difference between the numbers of the left brackets and the right brackets
+	int n,cnt=0,ans=0;// cnt is the difference between the numbers of the left brackets and the right brackets
 	scanf("%d%s",&n,&s);
 	for(int i=0,p=-1;i<n;++i)// p==-1 means no incorrect substring start postion was found, otherwise it contains the start position
 	{
@@ -16,6 +16,6 @@ int main()
 			p=-1;
 		}
 	}
-	printf("%ld",(cnt!=0)?-1:ans);// cnt!=0 means the numbers of left brackets and the right brackets are not equal
+	printf("%d",(cnt!=0)?-1:ans);// cnt!=0 means the numbers of left brackets and the right brackets are not equal
 	return 0;
 }
