+++ CF/1325/A/CF.1325A.0.cpp
@@ -5,9 +5,9 @@ int main()
 	int t;
 	for(scanf("%d",&t);t>0;--t)
 	{
-		long x;
-		scanf("%ld",&x);
-		printf("1 %ld\n",x-1);
+		int x;
+		scanf("%d",&x);
+		printf("1 %d\n",x-1);
 	}
 	return 0;
 }
